+++ src/core/config.json
