--- src/expenses-microservice/core/.env.development
@@ -1,9 +1,0 @@
-MONGO_DB_HOST=mongo-expense                            # when running the app without docker 
-MONGO_DB_DRIVER=mongodb
-API_SECRET=000                            # Used for creating a JWT. Can be anything 
-MONGO_DB_USER=mybudget
-MONGO_DB_PASSWORD=password
-MONGO_DB_NAME=monthlybudget
-MONGO_DB_PORT=27017
-
-APP_PORT=8081
