+++ src/model/expense.go
@@ -1,4 +1,4 @@
-package models
+package model
 
 import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
