+++ src/model/user.go
@@ -1,0 +1,1 @@
+package model
